fix videoautoplay test for IE
I made this as close as possible as the test in modernizr's master test
without breaking the fix to IE, this is as close as I got. There seems
to be 2 issues here, first is that neither of the videos that they are
using seems to work on IE (removing the extra video breaks IE) and
adding the videos without the source element seems to break it as well.

I'd like to merge this and try it for a while, if we can't find any
issues, I'll try to see if I can find a way to make it closer to their
test and get it merged.